# openapi.json
{"openapi":"3.1.0","info":{"title":"Auth API","description":"A simple auth API","license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"0.1.0"},"paths":{"/admin":{"get":{"tags":["protected"],"operationId":"admin_route","responses":{"200":{"description":"Admin access granted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"403":{"description":"Forbidden"}},"security":[{"api_key":[]}]}},"/login":{"post":{"tags":["auth"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Invalid credentials"}}}},"/register":{"post":{"tags":["auth"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"400":{"description":"Bad request"}}}}},"components":{"schemas":{"LoginRequest":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","required":["message","token"],"properties":{"message":{"type":"string"},"token":{"type":"string"}}},"RegisterRequest":{"type":"object","required":["first_name","last_name","email","password","confirm_password"],"properties":{"confirm_password":{"type":"string"},"email":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"},"password":{"type":"string"}}},"RegisterResponse":{"type":"object","required":["id","first_name","last_name","email"],"properties":{"email":{"type":"string"},"first_name":{"type":"string"},"id":{"type":"string"},"last_name":{"type":"string"}}},"Role":{"type":"string","enum":["Admin","User"]},"User":{"type":"object","required":["id","first_name","last_name","email","password","role"],"properties":{"email":{"type":"string"},"first_name":{"type":"string"},"id":{"type":"string"},"last_name":{"type":"string"},"password":{"type":"string"},"role":{"$ref":"#/components/schemas/Role"}}}}}}